+++ ChillaxScraps/Plugin.cs
@@ -15,7 +15,7 @@ namespace ChillaxScraps
     {
         const string GUID = "zigzag.chillaxscraps";
         const string NAME = "ChillaxScraps";
-        const string VERSION = "1.0.1";
+        const string VERSION = "1.1.0";
 
         public static Plugin instance;
         public static List<AudioClip> audioClips;
@@ -31,9 +31,11 @@ namespace ChillaxScraps
                 case 2: script = item.spawnPrefab.AddComponent<Boink>(); break;
                 case 3: script = item.spawnPrefab.AddComponent<Food>(); break;
                 case 4: script = item.spawnPrefab.AddComponent<UnoReverse>(); break;
+                case 5: script = item.spawnPrefab.AddComponent<EmergencyMeeting>(); break;
                 default: return;
             }
             script.grabbable = true;
+            script.isInFactory = true;
             script.grabbableToEnemies = true;
             script.itemProperties = item;
         }
@@ -48,7 +50,8 @@ namespace ChillaxScraps
             string directory = "Assets/Data/";
 
             gameObjects = new List<GameObject> {
-                bundle.LoadAsset<GameObject>(directory + "DeathNote/DeathNoteCanvas.prefab")
+                bundle.LoadAsset<GameObject>(directory + "DeathNote/DeathNoteCanvas.prefab"),
+                bundle.LoadAsset<GameObject>(directory + "EmergencyMeeting/EmergencyMeetingCanvas.prefab")
             };
 
             audioClips = new List<AudioClip> {
@@ -56,7 +59,8 @@ namespace ChillaxScraps
                 bundle.LoadAsset<AudioClip>(directory + "_audio/Death_Note_Heart_Attack_Sound_Effect.wav"),
                 bundle.LoadAsset<AudioClip>(directory + "_audio/Boink.wav"),
                 bundle.LoadAsset<AudioClip>(directory + "_audio/EatSFX.wav"),
-                bundle.LoadAsset<AudioClip>(directory + "_audio/uno-reverse-biaatch.wav")
+                bundle.LoadAsset<AudioClip>(directory + "_audio/uno-reverse-biaatch.wav"),
+                bundle.LoadAsset<AudioClip>(directory + "_audio/emergencymeeting.wav")
             };
 
             var scraps = new List<Scrap> {
@@ -66,7 +70,8 @@ namespace ChillaxScraps
                 new Scrap("CupNoodle/CupNoodleItem.asset", 11, 3),
                 new Scrap("Moai/MoaiItem.asset", 9),
                 new Scrap("UnoReverseCard/UnoReverseCardItem.asset", 8, 4),
-                new Scrap("FroggyChair/FroggyChairItem.asset", 10)
+                new Scrap("FroggyChair/FroggyChairItem.asset", 10),
+                new Scrap("EmergencyMeeting/EmergencyMeetingItem.asset", 6, 5)
             };
 
             int i = 0; config = new Config(base.Config, scraps);